+++ test/assets/pass/print.lua
@@ -17,4 +17,3 @@ print()
 
 -- Expect: 1
 print(1)
-